split __repr__ into draw and __repr__
`draw()` only draws the plot without showing it, `__repr__()` then uses
plt.show() to open the plot (which only does it if interactive == True).

To make this possible, all interactive = True has to be set to false
while in `draw()`. As usually every gui backend sets
interactive = True, we also don't need to set it in the theme.

Setting interactive = True results also in popping-up-windows when doing
unittests in nose.

This change is needed for saving a plot via `ggsave(plot)`.